Complete rewrite of the parser
The parser is now explicitly schema-driven, i.e., before we try to parse
an issue, it loads the schema from the server, and automatically parses
the fields according to that schema.

Client applications can provide overrides for custom fields, since the
type reported by the Jira server is not really reliable for them. Also,
the library itself has overrides for certain field types, as we cannot
really do anything with declared types like "array".